move index nested components into new file components
